Fix MSVC7 build issue
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@11431 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/btr.cpp
@@ -2594,7 +2594,7 @@ static SLONG fast_load(TDBB tdbb,
 		// how faster jumpkeys are recalculated on insert.
 		
 		
-		jumpInfo.jumpAreaSize = 512 + (sqrt(key_length) * 16);
+		jumpInfo.jumpAreaSize = 512 + ((int)sqrt((float)key_length) * 16);
 		//  key_size  |  jumpAreaSize
 		//  ----------+-----------------
 		//         4  |    544
